poprawilem funkcje checkRelationship

+++ PQL/src/QueryPreProcessor.cpp
@@ -26,7 +26,7 @@ void QueryPreProcessor::parseQuery(string query) {
 	vector<string> query_part;
 
 	Matcher* m = new Matcher();
-	for(int i = 0 ; i < first_split.size() ; i ++)
+	for(size_t i = 0 ; i < first_split.size() ; i ++)
 	{
 		if(!m->checkTokens(first_split[i], "select"))
 			result_part.push_back(first_split[i]);
